+++ Game/RainMeadow.GameHooks.cs
@@ -618,9 +618,9 @@ namespace RainMeadow
             {
                 self.cameras[0]?.hud.PlaySound(SoundID.MENY_Already_Selected_MultipleChoice_Clicked);
                 ChatTextBox.InvokeShutDownChat();
-                if ((self.cameras[0].hud.parts.Find(x => x is ChatHud) is ChatHud hud) && !hud.showChatLog)
+                if (RMOverlayHUDMenu.GetOverlay()?.chatHud is ChatHud chatHud && !chatHud.showChatLog)
                 {
-                    hud.ShutDownChatLog();
+                    chatHud.ShutDownChatLog();
                 }
             }
             if (OnlineManager.lobby != null)